voice-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
twilio>=8.0.0
openai>=1.0.0
httpx>=0.25.0
python-multipart>=0.0.6
